修复匹配错位的bug，添加数据长度为long

+++ myPort/界面/Form1.cs
@@ -562,7 +562,7 @@ namespace myPort
                 option.Series.TryRemove(str, out series);
             }
         }
-        private void chart_add_point(string str, int value)
+        private void chart_add_point(string str, long value)
         {
             if (!option.Series.ContainsKey(str))
             {
